mysite/requirements-dev.txt
Django==6.0.3
asgiref==3.11.1
sqlparse==0.5.5
pytest==9.0.2
pytest-django==4.12.0
factory-boy==3.3.3
faker==40.11.1
